fix: rename routers
