+++ src/components/Menu.tsx
@@ -31,7 +31,7 @@ export default function Menu() {
   const [, setToastText] = useAtom(toastTextAtom);
   const [, setToastType] = useAtom(toastTypeAtom);
 
-  const setWidgetLayoutMutation = api.widget.setWidgetLayout.useMutation({
+  const setWidgetLayoutMutation = api.layout.setWidgetLayout.useMutation({
     onSuccess: () => {
       setWidgetLayout(makeLayoutsStatic(editedWidgetLayout, true));
       setToastType("success");
