+++ src/server/api/routers/config.ts
@@ -10,30 +10,10 @@ import {
   updateUserWidgetLayout,
 } from "~/server/repository/layout/layoutRepository";
 import transformWidgetData from "~/server/service/transformWidgetDataService";
-import transformWidgetLayout from "~/server/service/transformWidgetLayoutService";
-import updateWidgetLayoutService from "~/server/service/updateWidgetLayoutService";
-import AppError from "~/utils/error";
 import Log from "~/utils/log";
-import {
-  screenSizePositioningSchema,
-  widgetLayoutSchema,
-  widgetTypeSchema,
-} from "~/utils/schema";
+import { screenSizePositioningSchema, widgetTypeSchema } from "~/utils/schema";
 
-export const widgetRouter = createTRPCRouter({
-  getWidgetLayout: protectedProcedure.query(async () => {
-    try {
-      const storedLayouts = await getLayoutRepository().getAll();
-      const layoutData = transformWidgetLayout(storedLayouts);
-      return layoutData;
-    } catch (error) {
-      Log(error, "error");
-      throw new TRPCError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: "Unable to get widget layout",
-      });
-    }
-  }),
+export const configRouter = createTRPCRouter({
   getWidgetConfig: protectedProcedure.query(async () => {
     try {
       const storedConfigs = await getConfigRepository().getAll();
@@ -47,25 +27,6 @@ export const widgetRouter = createTRPCRouter({
       });
     }
   }),
-  getLayoutForWidget: protectedProcedure
-    .input(z.object({ id: z.string() }))
-    .query(async ({ input }) => {
-      try {
-        const data = await getLayoutRepository().get(input.id);
-        if (!data) {
-          throw new AppError(
-            `No adjusted widget config for widget with ID ${input.id}`,
-          );
-        }
-        return data;
-      } catch (error) {
-        Log(error, "error");
-        throw new TRPCError({
-          code: "INTERNAL_SERVER_ERROR",
-          message: "Unable to get widget data",
-        });
-      }
-    }),
   setAllWidgetConfig: protectedProcedure
     .input(
       z.object({
@@ -105,28 +66,6 @@ export const widgetRouter = createTRPCRouter({
         throw new TRPCError({
           code: "INTERNAL_SERVER_ERROR",
           message: "Unable to update widget config",
-        });
-      }
-    }),
-  setWidgetLayout: protectedProcedure // TODO: rework
-    .input(
-      z.object({
-        layout: widgetLayoutSchema,
-      }),
-    )
-    .mutation(async ({ input }) => {
-      try {
-        const widgetConfig = await getLayoutRepository().get();
-        const updatedWidgets = updateWidgetLayoutService(
-          input.layout,
-          widgetConfig,
-        );
-        await getLayoutRepository().set(updatedWidgets);
-      } catch (error) {
-        Log(error, "error");
-        throw new TRPCError({
-          code: "INTERNAL_SERVER_ERROR",
-          message: "Unable to update widget layout",
         });
       }
     }),
