+++ src/server/api/routers/layout.ts
@@ -1,0 +1,66 @@
+import { TRPCError } from "@trpc/server";
+import { z } from "zod";
+import { createTRPCRouter, protectedProcedure } from "~/server/api/trpc";
+import { getLayoutRepository } from "~/server/repository/layout/layoutRepository";
+import transformWidgetLayout from "~/server/service/transformWidgetLayoutService";
+import updateWidgetLayoutService from "~/server/service/updateWidgetLayoutService";
+import AppError from "~/utils/error";
+import Log from "~/utils/log";
+import { widgetLayoutSchema } from "~/utils/schema";
+
+export const layoutRouter = createTRPCRouter({
+  getWidgetLayout: protectedProcedure.query(async () => {
+    try {
+      const storedLayouts = await getLayoutRepository().getAll();
+      const layoutData = transformWidgetLayout(storedLayouts);
+      return layoutData;
+    } catch (error) {
+      Log(error, "error");
+      throw new TRPCError({
+        code: "INTERNAL_SERVER_ERROR",
+        message: "Unable to get widget layout",
+      });
+    }
+  }),
+  getLayoutForWidget: protectedProcedure
+    .input(z.object({ id: z.string() }))
+    .query(async ({ input }) => {
+      try {
+        const data = await getLayoutRepository().get(input.id);
+        if (!data) {
+          throw new AppError(
+            `No adjusted widget config for widget with ID ${input.id}`,
+          );
+        }
+        return data;
+      } catch (error) {
+        Log(error, "error");
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Unable to get widget data",
+        });
+      }
+    }),
+  setWidgetLayout: protectedProcedure // TODO: rework
+    .input(
+      z.object({
+        layout: widgetLayoutSchema,
+      }),
+    )
+    .mutation(async ({ input }) => {
+      try {
+        const widgetConfig = await getLayoutRepository().get();
+        const updatedWidgets = updateWidgetLayoutService(
+          input.layout,
+          widgetConfig,
+        );
+        await getLayoutRepository().set(updatedWidgets);
+      } catch (error) {
+        Log(error, "error");
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Unable to update widget layout",
+        });
+      }
+    }),
+});
